Feat: Add a if1.sy case
change cg, to handle jumps correctly, instead of establishing a BB for each stmt

it is nonsense to generate a BB for each stmt. sequential stmt should be put into a contiguous BB.

it's just that we need to properly analyze BB-level RIG, and allocate registers accordingly.

remove redundant BB at the beginning of each function.